new commit

+++ src/main/resources/db/migration/V1__create_note_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE note (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    title VARCHAR(255),
-    content TEXT
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  title VARCHAR(255),
+  content TEXT
 );
